chore: extract Upcoming Departures into a LiveView
